requirements.txt
flask
requests
python-dotenv
flask
requests
python-dotenv
